Extract SupportedPaymentMethod to it's own top level class.
