+++ paymentsheet/src/main/java/com/stripe/android/customersheet/CustomerSheetState.kt
@@ -1,6 +1,6 @@
 package com.stripe.android.customersheet
 
-import com.stripe.android.lpmfoundations.luxe.LpmRepository
+import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.StripeIntent
 import com.stripe.android.paymentsheet.model.PaymentSelection
@@ -14,7 +14,7 @@ internal sealed interface CustomerSheetState {
         val config: CustomerSheet.Configuration?,
         val stripeIntent: StripeIntent?,
         val customerPaymentMethods: List<PaymentMethod>,
-        val supportedPaymentMethods: List<LpmRepository.SupportedPaymentMethod>,
+        val supportedPaymentMethods: List<SupportedPaymentMethod>,
         val isGooglePayReady: Boolean,
         val paymentSelection: PaymentSelection?,
         val cbcEligibility: CardBrandChoiceEligibility,
