+++ paymentsheet/src/main/java/com/stripe/android/customersheet/CustomerSheetViewAction.kt
@@ -1,6 +1,6 @@
 package com.stripe.android.customersheet
 
-import com.stripe.android.lpmfoundations.luxe.LpmRepository
+import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.payments.bankaccount.navigation.CollectBankAccountResultInternal
 import com.stripe.android.paymentsheet.forms.FormFieldValues
@@ -18,7 +18,7 @@ internal sealed class CustomerSheetViewAction {
     class OnModifyItem(val paymentMethod: PaymentMethod) : CustomerSheetViewAction()
     class OnItemRemoved(val paymentMethod: PaymentMethod) : CustomerSheetViewAction()
     class OnAddPaymentMethodItemChanged(
-        val paymentMethod: LpmRepository.SupportedPaymentMethod,
+        val paymentMethod: SupportedPaymentMethod,
     ) : CustomerSheetViewAction()
     class OnFormFieldValuesCompleted(
         val formFieldValues: FormFieldValues?,
