+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/luxe/LpmRepository.kt
@@ -1,9 +1,7 @@
 package com.stripe.android.lpmfoundations.luxe
 
 import android.content.res.Resources
-import androidx.annotation.DrawableRes
 import androidx.annotation.RestrictTo
-import androidx.annotation.StringRes
 import androidx.annotation.VisibleForTesting
 import com.stripe.android.model.LuxePostConfirmActionRepository
 import com.stripe.android.model.PaymentIntent
@@ -637,66 +635,6 @@ class LpmRepository(
             formSpec = LayoutSpec(sharedDataSpec.fields),
         )
         else -> null
-    }
-
-    /**
-     * Enum defining all payment method types for which Payment Sheet can collect
-     * payment data.
-     *
-     * FormSpec is optionally null only because Card is not converted to the
-     * compose model.
-     */
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    data class SupportedPaymentMethod(
-        /**
-         * This describes the PaymentMethod Type as described
-         * https://stripe.com/docs/api/payment_intents/create#create_payment_intent-payment_method_types
-         */
-        val code: PaymentMethodCode,
-
-        /** This describes if the LPM requires a mandate see [ConfirmPaymentIntentParams.mandateDataParams]. */
-        val requiresMandate: Boolean,
-
-        /** This describes the name that appears under the selector. */
-        @StringRes val displayNameResource: Int,
-
-        /** This describes the image in the LPM selector.  These can be found internally [here](https://www.figma.com/file/2b9r3CJbyeVAmKi1VHV2h9/Mobile-Payment-Element?node-id=1128%3A0) */
-        @DrawableRes val iconResource: Int,
-
-        /** An optional light theme icon url if it's supported. */
-        val lightThemeIconUrl: String?,
-
-        /** An optional dark theme icon url if it's supported. */
-        val darkThemeIconUrl: String?,
-
-        /** Indicates if the lpm icon in the selector is a single color and should be tinted
-         * on selection.
-         */
-        val tintIconOnSelection: Boolean,
-
-        /**
-         * This describes the requirements of the LPM including if it is supported with
-         * PaymentIntents w/ or w/out SetupFutureUsage set, SetupIntent, or on-session when attached
-         * to the customer object.
-         */
-        val requirement: PaymentMethodRequirements,
-
-        /**
-         * This describes how the UI should look.
-         */
-        val formSpec: LayoutSpec,
-
-        /**
-         * This forces the UI to render the required fields
-         */
-        val placeholderOverrideList: List<IdentifierSpec> = emptyList()
-    ) {
-        /**
-         * Returns true if the payment method supports confirming from a saved
-         * payment method of this type.  See [PaymentMethodRequirements] for
-         * description of the values
-         */
-        fun supportsCustomerSavedPM() = requirement.getConfirmPMFromCustomer(code)
     }
 
     @VisibleForTesting(otherwise = VisibleForTesting.PROTECTED)
