+++ paymentsheet/src/test/java/com/stripe/android/customersheet/utils/CustomerSheetTestHelper.kt
@@ -17,6 +17,7 @@ import com.stripe.android.customersheet.FakeCustomerAdapter
 import com.stripe.android.customersheet.FakeStripeRepository
 import com.stripe.android.customersheet.analytics.CustomerSheetEventReporter
 import com.stripe.android.lpmfoundations.luxe.LpmRepository
+import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodFixtures.CARD_PAYMENT_METHOD
 import com.stripe.android.networking.StripeRepository
@@ -178,7 +179,7 @@ internal object CustomerSheetTestHelper {
         ),
         isGooglePayAvailable: Boolean = true,
         customerPaymentMethods: List<PaymentMethod> = listOf(CARD_PAYMENT_METHOD),
-        supportedPaymentMethods: List<LpmRepository.SupportedPaymentMethod> = listOf(
+        supportedPaymentMethods: List<SupportedPaymentMethod> = listOf(
             LpmRepository.HardcodedCard,
             LpmRepository.hardCodedUsBankAccount,
         ),
